Added code
This is the VBA associated with a simple search form I built.  Our field
crews were using tablets with ArcMap and found the Find tool
overly-complicated for their needs.  This streamlined tool allows them
to search by the street name.  A results listbox shows the street and
cross street.  When they select a record in the listbox, the map zooms
to the block.